Initial work to split up circular deps
